refactor: rename domain/operation_history.rs to operation.rs
