+++ src/domain/mod.rs
@@ -1,4 +1,4 @@
-pub mod operation_history;
+pub mod operation;
 pub mod polyline;
 pub mod route;
 pub mod types;
